+++ app/src/main/java/com/example/practice/ui/main/First/RecyclerAdapter.java
@@ -23,6 +23,8 @@ import java.util.ArrayList;
 import com.example.practice.R;
 
 public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.ViewHolder> {
+    private final int cindex;
+    private final int cvindex;
     private ArrayList<Dictionary> mData;
     Context context;
 
@@ -46,7 +48,7 @@ public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.ViewHo
                 @Override
                 public void onClick(View v) {
                     AlertDialog.Builder builder = new AlertDialog.Builder(context);
-                    builder.setTitle("연락처").setMessage("("+group+") "+number);
+                    builder.setTitle("연락처"+cindex).setMessage(cvindex+"("+group+") "+number);
                     AlertDialog alertDialog = builder.create();
                     alertDialog.show();
                 }
@@ -123,9 +125,11 @@ public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.ViewHo
     }
 
     // 생성자에서 데이터 리스트 객체를 전달받음.
-    public RecyclerAdapter(ArrayList<Dictionary> list, Context context) {
+    public RecyclerAdapter(ArrayList<Dictionary> list, Context context, int cvindex, int cindex) {
         this.mData = list;
         this.context = context;
+        this.cvindex = cvindex;
+        this.cindex= cindex;
     }
