+++ app/src/main/res/drawable/ic_message_foreground.xml
@@ -1,11 +1,11 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="108dp"
     android:height="108dp"
-    android:viewportWidth="41.797283"
-    android:viewportHeight="41.797283"
+    android:viewportWidth="50.462574"
+    android:viewportHeight="50.462574"
     android:tint="#FFFFFF">
-  <group android:translateX="8.898642"
-      android:translateY="8.898642">
+  <group android:translateX="13.231287"
+      android:translateY="13.231287">
       <path
           android:fillColor="#FF000000"
           android:pathData="M20,2L4,2c-1.1,0 -1.99,0.9 -1.99,2L2,22l4,-4h14c1.1,0 2,-0.9 2,-2L22,4c0,-1.1 -0.9,-2 -2,-2zM18,14L6,14v-2h12v2zM18,11L6,11L6,9h12v2zM18,8L6,8L6,6h12v2z"/>
